app/requirements.txt
numpy==1.21.4
pandas==1.2.5
python-dateutil==2.8.2
pytz==2021.3
pywin32==300
PyYAML==5.4.1
six==1.16.0
XlsxWriter==1.3.9
pyperclip==1.8.2
